requirements.txt
flask
sympy
numpy
plotly
gunicorn
requests
mangum
asgirefflask
sympy
numpy
plotly
gunicorn
requests
